+++ setupMaster.sh
@@ -8,7 +8,8 @@ if [[ $EUID -ne 0 ]]; then
   exit 1
 fi
 
-GITHUB_URL="https://github.com/MrKaKisen/nafsdm.git"
+#  DL_VERSION will be changed at the time of update
+DL_URL="https://github.com/MrKaKisen/nafsdm/archive/"
 GITHUB_DIR="master-nafsdm"
 HOME_DIR="/home/master-nafsdm"
 USER="master-nafsdm"
@@ -22,7 +23,7 @@ read OPERATINGSYS
 if [ "$OPERATINGSYS" == "centos" ]; then
   echo "Installing packages.."
   yum update -y
-  yum install python git -y
+  yum install python curl wget -y
 
   # centos does not have pip in it's repos
   curl "https://bootstrap.pypa.io/get-pip.py" -o "get-pip.py"
@@ -33,7 +34,7 @@ if [ "$OPERATINGSYS" == "centos" ]; then
 elif [[ "$OPERATINGSYS" == "debian" ]] || [[ "$OPERATINGSYS" == "ubuntu" ]] ; then
   echo "Installing packages.."
   apt-get update -y
-  apt-get install python python-pip git -y
+  apt-get install python python-pip curl wget -y
 
   pip install requests
 else
@@ -41,12 +42,43 @@ else
   exit 1
 fi
 
+# get which version is the latest
+echo "Fetching information about latest version.."
+LATEST_VERSION=$(curl https://raw.githubusercontent.com/MrKaKisen/nafsdm/master/version.txt)
+
+# select version
+echo "Please select your version. Type in the version number or type 'latest' for latest version."
+echo -n "Version: "
+read VERSION_USER
+
+if [ "$VERSION_USER" == "latest" ]; then
+  echo -n "Confirm? (y/n): "
+  read CONFIRM
+  if [ "$CONFIRM" == "y" ]; then
+    DL_VERSION="$LATEST_VERSION"
+  else
+    echo "Aborting.."
+    exit 1
+  fi
+else
+  echo -n "Confirm? If version doesn't exist, script will fail. (y/n): "
+  read CONFIRM
+  if [ "$CONFIRM" == "y" ]; then
+    DL_VERSION="$VERSION_USER"
+  else
+    echo "Aborting.."
+    exit 1
+  fi
+fi
+
 echo "Required packages installed!"
 echo "Downloading nafsdm & installing.."
 
 # download in temp dir
 cd /tmp
-git clone $GITHUB_URL
+wget $DL_URL$DL_VERSION.tar.gz -O nafsdm.tar.gz
+tar -zxvf nafsdm.tar.gz
+mv nafsdm-* nafsdm
 
 useradd $USER
 # debian and ubuntu doesn't create its home dir automatically, unlike centos