+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version 1.1 (Status: RELEASED)
+* New version files.
+* New config parsing.
+* Setup scripts has new version selection (and another fix on that since 1.1-dev2)
+* Minor bug/spell fixes in master aswell as slave config changed (section name)
+
 Version 1.0.1-stable (Status: RELEASED)
 * Slave: Fixed serious writing issue where ' is used instead of ".
 * Slave: Updated hostname finding function.